main/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pika==1.2.0
sqlalchemy==1.4.23
pydantic==2.7.1
python-dotenv==1.0.1
pika==1.2.0
pymongo==4.7.2
motor==3.3.2
aio_pika==9.4.1
bcrypt==4.1.2